requirements.txt
bert_score
blobfile
nltk
numpy
packaging
psutil
PyYAML
setuptools
spacy
torchmetrics
tqdm
transformers
wandb
datasets
gradio
jittor